+++ NEWS.md
@@ -2,6 +2,9 @@
 
 * Better SuperTinyIcons shiny example
 * Removed jsonlite dependency
+* Allow to pass a list of options to the `contextMenu` argument
+* The value of `jstreeOutput` is now updated when the user 
+creates/deletes/pastes a node
 
 
 # jsTreeR 0.1.0 (2020-08-26)