+++ NEWS.md
@@ -4,7 +4,7 @@
 * Removed jsonlite dependency
 * Allow to pass a list of options to the `contextMenu` argument
 * The value of `jstreeOutput` is now updated when the user 
-creates/deletes/pastes a node
+creates/deletes/renames/pastes a node
 
 
 # jsTreeR 0.1.0 (2020-08-26)